+++ lib/install/jet.bash
@@ -13,7 +13,6 @@ elif [ -z "${_SHELL_GR_DIR:-}" ]; then
   _SHELL_GR_DIR="${_ROOT_DIR}"
 fi
 # Library Sourcing
-source "${_SHELL_GR_DIR}/lib/checksum.bash"              # verify_with_checksum_string_in_file
 source "${_SHELL_GR_DIR}/lib/debug.bash"                 # is_debug
 source "${_SHELL_GR_DIR}/lib/error.bash"                 # assert_not_empty, fail
 source "${_SHELL_GR_DIR}/lib/install/common/github.bash" # GRIC_GH_latest_stable, GRIC_GH_list_all_versions
@@ -23,32 +22,32 @@ source "${_SHELL_GR_DIR}/lib/temp.bash"                  # temp_dir
 source "${_SHELL_GR_DIR}/lib/trap.bash"                  # add_on_exit
 
 # shellcheck disable=SC2034
-GH_REPO="https://github.com/clj-kondo/clj-kondo"
-TOOL_NAME="clj-kondo"
-TOOL_TEST="clj-kondo --help"
+GH_REPO="https://github.com/borkdude/jet"
+TOOL_NAME="jet"
+TOOL_TEST="jet --help"
 
-GRI_CLJ_KONDO__list_deps() {
+GRI_JET__list_deps() {
   initial_deps=()
-  initial_deps+=(sort uniq)            # GRI_CLJ_KONDO__list_deps
-  initial_deps+=(curl unzip sha256sum) # GRI_CLJ_KONDO__download
-  initial_deps+=(git grep cut sed)     # GRIC_GH_list_github_tags
-  initial_deps+=(sed sort awk)         # GRIC_GH_sort_versions
-  initial_deps+=(curl sed tail xarg)   # GRIC_GH_latest_stable
+  initial_deps+=(sort uniq)          # GRI_JET__list_deps
+  initial_deps+=(curl tar)           # GRI_JET__download
+  initial_deps+=(git grep cut sed)   # GRIC_GH_list_github_tags
+  initial_deps+=(sed sort awk)       # GRIC_GH_sort_versions
+  initial_deps+=(curl sed tail xarg) # GRIC_GH_latest_stable
   mapfile -t deps < <(printf "%s\n" "${initial_deps[@]}" | sort | uniq)
   for item in "${deps[@]}"; do
     printf "%s\n" "${item}"
   done
 }
 
-GRI_CLJ_KONDO__list_all_versions() {
+GRI_JET__list_all_versions() {
   # inputs
   local -r gh_repo="${GH_REPO}"
   # body
   GRIC_GH_list_all_versions "${gh_repo}"
 }
-#GRI_CLJ_KONDO__list_all_versions
+#GRI_JET__list_all_versions
 
-GRI_CLJ_KONDO__latest_stable() {
+GRI_JET__latest_stable() {
   # inputs
   local -r gh_repo="${GH_REPO}"
   local -r github_api_token="${GITHUB_API_TOKEN:-}" # optional
@@ -56,9 +55,9 @@ GRI_CLJ_KONDO__latest_stable() {
   GITHUB_API_TOKEN="${github_api_token}" \
     GRIC_GH_latest_stable "${gh_repo}"
 }
-#GRI_CLJ_KONDO__latest_stable
+#GRI_JET__latest_stable
 
-GRI_CLJ_KONDO__compose_download_url() {
+GRI_JET__compose_download_url() {
   local version="$1"
   local platform_uname platform
   platform_uname="$(uname -s)"
@@ -84,19 +83,19 @@ GRI_CLJ_KONDO__compose_download_url() {
   # possible values:
   # https://stackoverflow.com/questions/45125516/possible-values-for-uname-m
 
-  # releases: https://github.com/clj-kondo/clj-kondo/releases
-  # sample URL: https://github.com/clj-kondo/clj-kondo/releases/download/v2023.12.15/clj-kondo-2023.12.15-linux-amd64.zip
-  local download_url="${GH_REPO}/releases/download/v${version}/${TOOL_NAME}-${version}-${platform}-${arch}.zip"
+  # releases: https://github.com/borkdude/jet/releases
+  # sample URL: https://github.com/borkdude/jet/releases/download/v0.7.27/jet-0.7.27-linux-amd64.tar.gz
+  local download_url="${GH_REPO}/releases/download/v${version}/${TOOL_NAME}-${version}-${platform}-${arch}.tar.gz"
   printf "%s" "${download_url}"
 }
-#GRI_CLJ_KONDO__compose_download_url "2024.05.24"
+#GRI_JET__compose_download_url "0.7.27"
 
-GRI_CLJ_KONDO__download() {
+GRI_JET__download() {
   # inputs
-  local -r type="${GRI_CLJ_KONDO__INSTALL_TYPE:-}"
-  local -r version="${GRI_CLJ_KONDO__INSTALL_VERSION:-}"
-  local -r dest="${GRI_CLJ_KONDO__DOWNLOAD_PATH:-}"
-  local -r github_api_token="${GRI_CLJ_KONDO__GITHUB_API_TOKEN:-}" # optional
+  local -r type="${GRI_JET__INSTALL_TYPE:-}"
+  local -r version="${GRI_JET__INSTALL_VERSION:-}"
+  local -r dest="${GRI_JET__DOWNLOAD_PATH:-}"
+  local -r github_api_token="${GRI_JET__GITHUB_API_TOKEN:-}" # optional
 
   # inputs validation
   [ "${type}" != "version" ] && fail "Only installation by version is supported."
@@ -111,47 +110,44 @@ GRI_CLJ_KONDO__download() {
   log_debug
 
   # prepare curl opts
-  local curl_opts=(-fsSL)
+  local curl_opts=(
+    -L #  follow redirects
+  )
   if [ -n "${github_api_token:-}" ]; then
     curl_opts=("${curl_opts[@]}" "-H" "Authorization: token ${github_api_token}")
   fi
 
   # download archive & checksum file
   log_info "Downloading ${TOOL_NAME} release ${version}..."
-  local archive_url checksum_url
-  archive_url="$(GRI_CLJ_KONDO__compose_download_url "${version}")"
-  checksum_url="${archive_url}.sha256"
+  local archive_url
+  archive_url="$(GRI_JET__compose_download_url "${version}")"
   log_info " - ${archive_url}"
-  log_info " - ${checksum_url}"
-  local temp_archive_path="${temp_dir}/${TOOL_NAME}.zip"
-  local temp_checksum_path="${temp_dir}/${TOOL_NAME}.zip.sha256"
+  local temp_archive_path="${temp_dir}/${TOOL_NAME}.tar.gz"
   curl "${curl_opts[@]}" -o "${temp_archive_path}" -C - "${archive_url}" || fail "Could not download ${archive_url}"
-  curl "${curl_opts[@]}" -o "${temp_checksum_path}" -C - "${checksum_url}" || fail "Could not download ${checksum_url}"
-  log_info
-
-  # verify checksum
-  GR_CHECKSUM__ALGO="sha256" \
-    GR_CHECKSUM__FILE_PATH="${temp_archive_path}" \
-    GR_CHECKSUM__CHECKSUM_PATH="${temp_checksum_path}" \
-    verify_with_checksum_string_in_file
   log_info
 
   # move the downloaded file to final destination
   mkdir -p "${dest}"
-  local unzip_opts=(-o "${temp_archive_path}" -d "${dest}")
-  ! is_debug && unzip_opts=(-q "${unzip_opts[@]}")
-  unzip "${unzip_opts[@]}" || fail "Could not extract ${temp_archive_path}"
+  local tar_opts=(-xzf "${temp_archive_path}" -C "${dest}")
+  is_debug && tar_opts+=(-v)
+  tar "${tar_opts[@]}" || fail "Could not extract ${temp_archive_path}"
 
   log_info "Downloading ${TOOL_NAME} release ${version}... DONE"
   log_info
 }
+#rm -rf /tmp/tmp-jet-download-dir
+#GRI_JET__INSTALL_TYPE="version" \
+#  GRI_JET__INSTALL_VERSION="0.7.27" \
+#  GRI_JET__DOWNLOAD_PATH="/tmp/tmp-jet-download-dir" \
+#  GRI_JET__GITHUB_API_TOKEN="$(pass show rynkowski/github/rynkowsg/token/release_download_token)" \
+#  GRI_JET__download
 
-GRI_CLJ_KONDO__install_downloaded() {
+GRI_JET__install_downloaded() {
   # inputs
-  local -r type="${GRI_CLJ_KONDO__INSTALL_TYPE:-}"
-  local -r version="${GRI_CLJ_KONDO__INSTALL_VERSION:-}"
-  local -r download_path="${GRI_CLJ_KONDO__DOWNLOAD_PATH:-}"
-  local -r install_path="${GRI_CLJ_KONDO__INSTALL_PATH:-}"
+  local -r type="${GRI_JET__INSTALL_TYPE:-}"
+  local -r version="${GRI_JET__INSTALL_VERSION:-}"
+  local -r download_path="${GRI_JET__DOWNLOAD_PATH:-}"
+  local -r install_path="${GRI_JET__INSTALL_PATH:-}"
 
   # inputs validation
   [ "${type}" != "version" ] && fail "Only installation by version is supported."
@@ -174,16 +170,23 @@ GRI_CLJ_KONDO__install_downloaded() {
     fail "An error occurred while installing ${TOOL_NAME} ${version}."
   )
 }
+#rm -rf /tmp/tmp-jet-install-dir
+#GRI_JET__INSTALL_TYPE="version" \
+#  GRI_JET__INSTALL_VERSION="0.7.27" \
+#  GRI_JET__DOWNLOAD_PATH="/tmp/tmp-jet-download-dir" \
+#  GRI_JET__INSTALL_PATH="/tmp/tmp-jet-install-dir" \
+#  GRI_JET__install_downloaded
+#/tmp/tmp-jet-install-dir/jet --version
 
-# Installs clj-kondo using the specified version.
+# Installs jet using the specified version.
 # Basically it runs under the hood two functions:
-# - GRI_CLJ_KONDO__download &
-# - GRI_CLJ_KONDO__install_downloaded
-GRI_CLJ_KONDO__install() {
+# - GRI_JET__download &
+# - GRI_JET__install_downloaded
+GRI_JET__install() {
   # inputs
-  local -r type="${GRI_CLJ_KONDO__INSTALL_TYPE:-}"
-  local -r version="${GRI_CLJ_KONDO__INSTALL_VERSION:-}"
-  local -r install_path="${GRI_CLJ_KONDO__INSTALL_PATH:-}"
+  local -r type="${GRI_JET__INSTALL_TYPE:-}"
+  local -r version="${GRI_JET__INSTALL_VERSION:-}"
+  local -r install_path="${GRI_JET__INSTALL_PATH:-}"
   local -r github_api_token="${GITHUB_API_TOKEN:-}" # optional
 
   # inputs validation
@@ -196,15 +199,15 @@ GRI_CLJ_KONDO__install() {
   log_debug "Directory to store downloaded files created at ${download_path}"
   ! is_debug && add_on_exit rm -rf "${download_path}"
 
-  GRI_CLJ_KONDO__INSTALL_TYPE="${type}" \
-    GRI_CLJ_KONDO__INSTALL_VERSION="${version}" \
-    GRI_CLJ_KONDO__DOWNLOAD_PATH="${download_path}" \
+  GRI_JET__INSTALL_TYPE="${type}" \
+    GRI_JET__INSTALL_VERSION="${version}" \
+    GRI_JET__DOWNLOAD_PATH="${download_path}" \
     GITHUB_API_TOKEN="${github_api_token}" \
-    GRI_CLJ_KONDO__download
+    GRI_JET__download
 
-  GRI_CLJ_KONDO__INSTALL_TYPE="${type}" \
-    GRI_CLJ_KONDO__INSTALL_VERSION="${version}" \
-    GRI_CLJ_KONDO__INSTALL_PATH="${install_path}" \
-    GRI_CLJ_KONDO__DOWNLOAD_PATH="${download_path}" \
-    GRI_CLJ_KONDO__install_downloaded
+  GRI_JET__INSTALL_TYPE="${type}" \
+    GRI_JET__INSTALL_VERSION="${version}" \
+    GRI_JET__INSTALL_PATH="${install_path}" \
+    GRI_JET__DOWNLOAD_PATH="${download_path}" \
+    GRI_JET__install_downloaded
 }